sendmail/requirements.txt
google-cloud-storage
google-cloud-pubsub
Flask-Mail
flask
